requirements.txt
streamlit
pandas
matplotlib
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
PyPDF2
tiktoken
